feat: navigation setup finished
