+++ app/src/main/java/com/example/bellascoffeeapp/composables/screens/Home.kt
@@ -1,13 +1,31 @@
 package com.example.bellascoffeeapp.composables.screens
 
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import com.example.bellascoffeeapp.ShopCoffeeBags
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import com.example.bellascoffeeapp.ui.theme.BellasTheme
 import kotlinx.serialization.Serializable
 
 @Composable
 fun Home(){
-    ShopCoffeeBags()
+    Column (
+        modifier = Modifier.fillMaxSize()
+        .background(BellasTheme.colorScheme.background),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ){
+
+            Text(text = "Home page",
+            style = BellasTheme.typography.titleMedium)
+    }
 }
+
+
 
 @Serializable
 object Home
