+++ app/src/main/java/com/example/bellascoffeeapp/composables/screens/Shop.kt
@@ -1,12 +1,12 @@
 package com.example.bellascoffeeapp.composables.screens
 
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import com.example.bellascoffeeapp.composables.ShopCoffeeBags
 import kotlinx.serialization.Serializable
 
 @Composable
 fun Shop(){
-    Text(text = "Shop page")
+    ShopCoffeeBags()
 }
 
 @Serializable
