--- app/src/main/java/com/example/bellascoffeeapp/ui/theme/Color.kt
@@ -1,3 +1,0 @@
-package com.example.bellascoffeeapp.ui.theme
-
-import androidx.compose.ui.graphics.Color
