+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">BellasCoffeeApp</string>
 
     <string name="bcl">BELLAS COFFEE LAB</string>
+
 </resources>
